Created tasks, shot tasks, notes, and employee hours tables.

+++ data/migrations/20190702202539_create_employees_table.js
@@ -6,7 +6,7 @@ exports.up = function(knex, Promise) {
       .primary();
 
     tbl
-      .integer("rate")
+      .decimal("rate")
       .unsigned()
       .defaultTo(null);
